requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blobfile==3.0.0
build==1.2.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
cryptography==43.0.1
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.9
distro==1.9.0
docutils==0.21.2
executing==2.1.0
fastapi==0.115.0
fastjsonschema==2.20.0
filelock==3.16.1
fire==0.6.0
flake8==7.1.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
gotrue==2.9.2
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.27.0
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==25.4.1
lazy_loader==0.4
llama_models==0.0.36
llama_stack==0.0.36
lxml==5.3.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.5.0
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nh3==0.2.18
notebook_shim==0.2.4
numpy==2.1.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pipdeptree==2.16.2
pkginfo==1.10.0
platformdirs==4.3.6
postgrest==0.17.1
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.18.0
pyproject_hooks==1.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-json-logger==2.0.7
python-multipart==0.0.11
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
readme_renderer==44.0
realtime==2.0.5
redis==5.1.0
referencing==0.35.1
reflex==0.6.1
reflex-chakra==0.6.0
reflex-hosting-cli==0.1.13
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.20.0
SecretStorage==3.3.3
selenium==4.25.0
Send2Trash==1.8.3
setuptools==70.1.1
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.35
sqlmodel==0.0.22
stack-data==0.6.3
starlette==0.38.6
starlette-admin==0.14.1
storage3==0.8.1
StrEnum==0.4.15
supabase==2.9.0
supafunc==0.6.1
tabulate==0.9.0
termcolor==2.4.0
terminado==0.18.1
tiktoken==0.7.0
tinycss2==1.3.0
tomlkit==0.13.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
twine==5.1.1
typer==0.12.5
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.2
undetected-chromedriver==3.5.5
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.31.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.13.1
zipp==3.20.2
